OPENE-368 Addressing code review remarks.
